+++ java/java.lsp.server/nbproject/org-netbeans-modules-java-lsp-server.sig
@@ -1,5 +1,5 @@
 #Signature file v4.1
-#Version 1.10.0
+#Version 1.14.0
 
 CLSS public java.lang.Object
 cons public init()
@@ -22,6 +22,35 @@ meth public java.lang.Object notify(org.openide.NotifyDescriptor)
 meth public void notifyLater(org.openide.NotifyDescriptor)
 supr org.openide.DialogDisplayer
 hfds context
+
+CLSS public org.netbeans.modules.java.lsp.server.ui.AbstractGlobalActionContext
+cons public init()
+intf org.openide.util.ContextGlobalProvider
+intf org.openide.util.Lookup$Provider
+meth public org.openide.util.Lookup createGlobalContext()
+meth public org.openide.util.Lookup getLookup()
+meth public static <%0 extends java.lang.Object> {%%0} withActionContext(org.openide.util.Lookup,java.util.concurrent.Callable<{%%0}>)
+supr java.lang.Object
+hcls ContextHolder
+
+CLSS public org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer
+cons protected init()
+innr protected final View
+intf org.netbeans.spi.htmlui.HTMLViewerSpi<org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View,java.lang.Object>
+meth public <%0 extends java.lang.Object> {%%0} component(org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View,java.lang.Class<{%%0}>)
+meth public java.lang.Object createButton(org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View,java.lang.String)
+meth public java.lang.String getId(org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View,java.lang.Object)
+meth public org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View newView(org.netbeans.spi.htmlui.HTMLViewerSpi$Context)
+meth public void runLater(org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View,java.lang.Runnable)
+meth public void setEnabled(org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View,java.lang.Object,boolean)
+meth public void setText(org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View,java.lang.Object,java.lang.String)
+supr java.lang.Object
+hfds initial
+
+CLSS protected final org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer$View
+ outer org.netbeans.modules.java.lsp.server.ui.AbstractLspHtmlViewer
+supr java.lang.Object
+hfds ctx,presenter,ui
 
 CLSS public abstract org.netbeans.modules.java.lsp.server.ui.AbstractLspInputOutputProvider
 cons protected init()
@@ -92,6 +121,7 @@ meth protected abstract java.util.concurrent.CompletableFuture<org.eclipse.lsp4j
 meth protected abstract org.openide.awt.StatusDisplayer$Message showStatusMessage(org.netbeans.modules.java.lsp.server.protocol.ShowStatusMessageParams)
 meth protected abstract void logMessage(org.eclipse.lsp4j.MessageParams)
 meth protected abstract void showMessage(org.eclipse.lsp4j.MessageParams)
+meth protected java.util.concurrent.CompletableFuture<java.lang.String> showHtmlPage(org.netbeans.modules.java.lsp.server.protocol.HtmlPageParams)
 meth public static org.netbeans.modules.java.lsp.server.ui.UIContext find()
  anno 0 org.netbeans.api.annotations.common.NonNull()
 meth public static org.netbeans.modules.java.lsp.server.ui.UIContext find(org.openide.util.Lookup)
@@ -103,6 +133,16 @@ hcls LogImpl
 CLSS public abstract interface org.netbeans.modules.progress.spi.ProgressEnvironment
 meth public abstract org.netbeans.api.progress.ProgressHandle createHandle(java.lang.String,org.openide.util.Cancellable,boolean)
 meth public abstract org.netbeans.modules.progress.spi.Controller getController()
+
+CLSS public abstract interface org.netbeans.spi.htmlui.HTMLViewerSpi<%0 extends java.lang.Object, %1 extends java.lang.Object>
+innr public final static Context
+meth public abstract <%0 extends java.lang.Object> {%%0} component({org.netbeans.spi.htmlui.HTMLViewerSpi%0},java.lang.Class<{%%0}>)
+meth public abstract java.lang.String getId({org.netbeans.spi.htmlui.HTMLViewerSpi%0},{org.netbeans.spi.htmlui.HTMLViewerSpi%1})
+meth public abstract void runLater({org.netbeans.spi.htmlui.HTMLViewerSpi%0},java.lang.Runnable)
+meth public abstract void setEnabled({org.netbeans.spi.htmlui.HTMLViewerSpi%0},{org.netbeans.spi.htmlui.HTMLViewerSpi%1},boolean)
+meth public abstract void setText({org.netbeans.spi.htmlui.HTMLViewerSpi%0},{org.netbeans.spi.htmlui.HTMLViewerSpi%1},java.lang.String)
+meth public abstract {org.netbeans.spi.htmlui.HTMLViewerSpi%0} newView(org.netbeans.spi.htmlui.HTMLViewerSpi$Context)
+meth public abstract {org.netbeans.spi.htmlui.HTMLViewerSpi%1} createButton({org.netbeans.spi.htmlui.HTMLViewerSpi%0},java.lang.String)
 
 CLSS public abstract interface org.netbeans.spi.io.InputOutputProvider<%0 extends java.lang.Object, %1 extends java.io.PrintWriter, %2 extends java.lang.Object, %3 extends java.lang.Object>
 meth public abstract boolean isIOClosed({org.netbeans.spi.io.InputOutputProvider%0})
@@ -189,4 +229,28 @@ meth public static org.openide.awt.StatusDisplayer getDefault()
 supr java.lang.Object
 hfds INSTANCE
 hcls Trivial
+
+CLSS public abstract interface org.openide.util.ContextGlobalProvider
+meth public abstract org.openide.util.Lookup createGlobalContext()
+
+CLSS public abstract org.openide.util.Lookup
+cons public init()
+fld public final static org.openide.util.Lookup EMPTY
+innr public abstract interface static Provider
+innr public abstract static Item
+innr public abstract static Result
+innr public final static Template
+meth public <%0 extends java.lang.Object> java.util.Collection<? extends {%%0}> lookupAll(java.lang.Class<{%%0}>)
+meth public <%0 extends java.lang.Object> org.openide.util.Lookup$Item<{%%0}> lookupItem(org.openide.util.Lookup$Template<{%%0}>)
+meth public <%0 extends java.lang.Object> org.openide.util.Lookup$Result<{%%0}> lookupResult(java.lang.Class<{%%0}>)
+meth public abstract <%0 extends java.lang.Object> org.openide.util.Lookup$Result<{%%0}> lookup(org.openide.util.Lookup$Template<{%%0}>)
+meth public abstract <%0 extends java.lang.Object> {%%0} lookup(java.lang.Class<{%%0}>)
+meth public static org.openide.util.Lookup getDefault()
+supr java.lang.Object
+hfds LOG,defaultLookup,defaultLookupProvider
+hcls DefLookup,Empty
+
+CLSS public abstract interface static org.openide.util.Lookup$Provider
+ outer org.openide.util.Lookup
+meth public abstract org.openide.util.Lookup getLookup()
